exam-recognition/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
python-docx>=0.8.11
baidu-aip>=2.2.18.0
Pillow>=9.0.0
Werkzeug>=2.0.0
